back/requirements.txt
Django~=4.2.13
djangorestframework~=3.15.1
pip~=24.0
setuptools~=70.0.0
django-cors-headers~=4.0.0
drf-yasg~=1.21.0
django-serializer~=1.2.1
psycopg2-binary~=2.9.3
whitenoise~=6.6.0
pypdf2~=3.0.1
transformers~=4.41.2
wheel~=0.37.0